Remove template option, as it was stale and unused.
