--- src/Templates/Android.md
@@ -1,112 +1,0 @@
-﻿id:XXX
-title:Xamarin.Android XXX
-clio-start-range:{OLDEST_COMMIT_CONSIDERED}
-clio-end-range:{NEWEST_COMMIT_CONSIDERED}
-clio-include-starting:{INCLUDE_OLDEST}
-
-Xamarin.Android XXX: XXX
-
-# System Requirements
-
-Xamarin.Android XXX requires
-[JDK 1.8](http://www.oracle.com/technetwork/java/javase/downloads/jdk8-downloads-2133151.html)
-to use the Android Nougat (API 24) APIs. Using the Android build-tools r24+ package may also
-require using JDK 1.8.
-You can continue to use earlier versions of the JDK when using earlier build-tools packages
-and when targeting earlier Android API levels:
-
-* [JDK 1.8](http://www.oracle.com/technetwork/java/javase/downloads/jdk8-downloads-2133151.html) - up to API 24+
-
-* [JDK 1.7](http://www.oracle.com/technetwork/java/javase/downloads/jdk7-downloads-1880260.html) - up to API 23
-
-* [JDK 1.6](http://www.oracle.com/technetwork/java/javase/downloads/java-archive-downloads-javase6-419409.html) - up to API 20
-
-Additionally, a 64-bit version of the JDK is required to use
-[custom controls in the Android designer][gui-designer].
-
-[gui-designer]: https://developer.xamarin.com/releases/vs/xamarin.vs_4/xamarin.vs_4.2/#androiddesignercustomcontrols
-
-**The simplest option is to install the 64-bit version of JDK 1.8 since it is
-backwards compatible with all of the previous API levels and supports the new
-Android designer features.**
-
-
-# XXX Features
-
-
-# Known Issues
-
-XXX
-
-# Xamarin.Android XXX
-
-XXX
-
-## New Features
-
-XXX
-
-<a name="experimental-features"/>
-
-## Experimental Features
-
-XXX
-
-<a name="Breaking_Changes" />
-
-## Breaking Changes
-
-XXX
-
-
-<a name="Bug_Fixes"/>
-
-## Bug Fixes
-
-* [12935](https://bugzilla.xamarin.com/show_bug.cgi?id=12935):
-    sensorPortrait should be sensorPortait
-
-{BUG_LIST}
-
-
-<a name="integrated-mono" />
-
-## Integrated Mono Features/Fixes
-
-Xamarin.Android uses [Mono XXX](http://www.mono-project.com/docs/about-mono/releases/XXX/)
-[commit XXX](https://github.com/mono/mono/commit/XXX).
-
-
-XXX
-
-
-<a name="API_Changes" />
-
-# API Changes
-
-* API Level 10:
-    [Mono.Android.dll](level_10_diff/mono.android.dll),
-    [OpenTK.dll](level_10_diff/opentk.dll),
-    [OpenTK-1.0.dll](level_10_diff/opentk-1.0.dll)
-* API Level 15:
-    [Mono.Android.dll](level_15_diff/mono.android.dll)
-* API Level 16:
-    [Mono.Android.dll](level_16_diff/mono.android.dll)
-* API Level 17:
-    [Mono.Android.dll](level_17_diff/mono.android.dll)
-* API Level 18:
-    [Mono.Android.dll](level_18_diff/mono.android.dll)
-* API Level 19:
-    [Mono.Android.dll](level_19_diff/mono.android.dll)
-* API Level 20:
-    [Mono.Android.dll](level_20_diff/mono.android.dll)
-* API Level 21:
-    [Mono.Android.dll](level_21_diff/mono.android.dll)
-* API Level 22:
-    [Mono.Android.dll](level_22_diff/mono.android.dll)
-* API Level 23:
-    [Mono.Android.dll](level_23_diff/mono.android.dll)
-* API Level 24:
-    [Mono.Android.dll](level_24_diff/mono.android.dll)
-* API Level 25:
-    [Mono.Android.dll](level_25_diff/mono.android.dll)
