--- src/Templates/Mac.md
@@ -1,41 +1,0 @@
-﻿id:XXX
-title:Xamarin.Mac
-version:XXX
-releasedate:XXX
-clio-start-range:{OLDEST_COMMIT_CONSIDERED}
-clio-end-range:{NEWEST_COMMIT_CONSIDERED}
-clio-include-starting:{INCLUDE_OLDEST}
-
-Requirements
-============
-
-- The latest features and API requires Xcode XXX and the bundled macOS SDK;
-- Apple Xcode XXX requires a Mac running OSX XXX (XXX) or newer;
-
-What's New
-==========
-
-This release is built upon our [open sourced SDK](https://github.com/xamarin/xamarin-macios),
-using the `XXX` branch, and include some additional IDE integration tools.
-
-### XXX Features
-
-### Miscellaneous Enhancements
-
-**Tools**
-
-XXX
-
-**API**
-
-XXX
-
-### Bug Fixes
-
-{BUG_LIST}
-
-### API diff
-
-The following documents contains a complete list of the API changes since our latest stable release: XM XXX (XXX)
-
-* [macOS](/releases/mac/api_changes/XXX);
