--- src/Templates/iOS.md
@@ -1,43 +1,0 @@
-﻿id:XXX
-title:Xamarin.iOS
-version:XXX
-releasedate:XXX
-clio-start-range:{OLDEST_COMMIT_CONSIDERED}
-clio-end-range:{NEWEST_COMMIT_CONSIDERED}
-clio-include-starting:{INCLUDE_OLDEST}
-
-Requirements
-============
-
-- The latest features and API requires Xcode XXX and the bundled iOS, tvOS and watchOS SDKs;
-- Apple Xcode XXX requires a Mac running OSX XXX (XXX) or newer;
-
-What's New
-==========
-
-This release is built upon our [open sourced SDK](https://github.com/xamarin/xamarin-macios),
-using the `XXX` branch, and is based on Mono XXX.
-
-### XXX Features
-
-### Miscellaneous Enhancements
-
-**Tools**
-
-XXX
-
-**API**
-
-XXX
-
-### Bug Fixes
-
-{BUG_LIST}
-
-### API diff
-
-The following documents contains a complete list of the API changes since our previous stable 10.10 release:
-
-* [iOS](/releases/ios/api_changes/XXX);
-* [tvOS](/releases/ios/api_changes/XXX);
-* [watchOS](/releases/ios/api_changes/XXX)
