+++ src/clio.csproj
@@ -58,7 +58,6 @@
     <Compile Include="SearchOptions.cs" />
     <Compile Include="BugCollector.cs" />
     <Compile Include="Model\BugEntry.cs" />
-    <Compile Include="TemplateGenerator.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
@@ -66,18 +65,12 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Model\" />
-    <Folder Include="Templates\" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\external\libbz\libbz\libbz.csproj">
       <Project>{DB06A55C-89FF-46A9-ABE5-D9F095C6B740}</Project>
       <Name>libbz</Name>
     </ProjectReference>
-  </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Include="Templates\Mac.md" />
-    <EmbeddedResource Include="Templates\iOS.md" />
-    <EmbeddedResource Include="Templates\Android.md" />
   </ItemGroup>
   <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
 </Project>
